Code shuffle: raftServer in server.go, Raft in raft.go
After this change:
 - raft.go consists of the public-facing API exported by the library (was mostly in api.go before).
 - server.go consists of the raftServer struct and its methods (was mostly in raft.go before),
   as well as the raftShared struct (was in state.go before).

Towards #172: "Merge state.go into raft.go?"
